+++ .github/workflows/update_dashboard.yml
@@ -1,4 +1,4 @@
-name: Update dashboard
+name: Update Covid 19 dashboard
 
 on:
   # manual for now
@@ -9,18 +9,22 @@ on:
 
 
 jobs:
-  pipeline:
+  update_dashboard:
+    name: Update dashboard data repository
     runs-on: ubuntu-latest
     steps:
       - name: Check out etl-pipeline repo
         uses: actions/checkout@v2
+
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
           python-version: 3.8
+
       - name: Install Python dependencies
         run: |
           pip install -r requirements.txt
+
       - name: Execuiting UpdateGSheetTask
         run: |
           python -m luigi --module tasks UpdateGSheetTask --local-scheduler
