+++ backend/metrics/calculations.py
@@ -1,8 +1,6 @@
-from datetime import datetime
 import numpy as np
 import pandas as pd
 from numpy import random
-
 from scipy.interpolate import CubicSpline
