+++ tests/integration_tests/README.md
@@ -1,0 +1,4 @@
+# Integration test
+
+This is an other set of integration tests. All tests are small and focus on a small action and if possible should also
+be executed as part of code quality checks
