+++ tests_integration/README.md
@@ -1,7 +1,6 @@
 # Integration tests
 
-These tests are intended for manual run's<br />
+Unlike the integration tests under tests/integration_tests these tests are intended for manual run's<br />
 Mostly for testing and debugging integration issues of the pipeline The reason you don't want to run them every time
 because some of them query external sources like API's. Others take a significant amount of time to run.
 
-Also good to use this on preping a release.
